MRParallelMinMax.h
