+++ source/MRIOExtras/MRTiff.cpp
@@ -5,7 +5,7 @@
 #include "MRMesh/MRChunkIterator.h"
 #include "MRMesh/MRFinally.h"
 #include "MRMesh/MRIOFormatsRegistry.h"
-#include "MRMesh/MRParallelFor.h"
+#include "MRMesh/MRParallelMinMax.h"
 #include "MRMesh/MRStringConvert.h"
 #include "MRPch/MRFmt.h"
