+++ source/MRVoxels/MRToolPath.cpp
@@ -8,7 +8,7 @@
 #include "MRMesh/MRExtractIsolines.h"
 #include "MRMesh/MRSurfaceDistance.h"
 #include "MRMesh/MRMeshDirMax.h"
-#include "MRMesh/MRParallelFor.h"
+#include "MRMesh/MRParallelMinMax.h"
 #include "MRMesh/MRObjectGcode.h"
 #include "MRMesh/MRExpected.h"
 #include "MRMesh/MRMeshIntersect.h"
@@ -24,8 +24,6 @@
 #include "MRMesh/MRInnerShell.h"
 #include "MRMesh/MRRingIterator.h"
 
-#include "MRMesh/MRParallelFor.h"
-#include "MRPch/MRTBB.h"
 #include <sstream>
 #include <span>
